VueComponents: Move SortPanel to "listing" directory
